# contracts/openapi/v1/booking.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "SportHub Booking",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/groups": {
      "get": {
        "operationId": "listMyGroups",
        "tags": [
          "Group Booking"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Group"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Membership, owner, payer or signed invite mismatch"
          },
          "409": {
            "description": "Deadline, slot, allocation or isolated BLOCKED_RULE conflict"
          }
        }
      },
      "post": {
        "operationId": "createGroupBooking",
        "tags": [
          "Group Booking"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Successful result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Group"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Membership, owner, payer or signed invite mismatch"
          },
          "409": {
            "description": "Deadline, slot, allocation or isolated BLOCKED_RULE conflict"
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroup"
              }
            }
          }
        }
      }
    },
    "/groups/{id}": {
      "get": {
        "operationId": "getGroupBooking",
        "tags": [
          "Group Booking"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Group"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Membership, owner, payer or signed invite mismatch"
          },
          "409": {
            "description": "Deadline, slot, allocation or isolated BLOCKED_RULE conflict"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/groups/{id}/invite": {
      "get": {
        "operationId": "getGroupInvitation",
        "tags": [
          "Group Booking"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/GroupInvite"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Membership, owner, payer or signed invite mismatch"
          },
          "409": {
            "description": "Deadline, slot, allocation or isolated BLOCKED_RULE conflict"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/groups/invitations": {
      "get": {
        "operationId": "previewGroupInvitation",
        "tags": [
          "Group Booking"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Group"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Membership, owner, payer or signed invite mismatch"
          },
          "409": {
            "description": "Deadline, slot, allocation or isolated BLOCKED_RULE conflict"
          }
        },
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/groups/join": {
      "post": {
        "operationId": "joinGroupBooking",
        "tags": [
          "Group Booking"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Group"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Membership, owner, payer or signed invite mismatch"
          },
          "409": {
            "description": "Deadline, slot, allocation or isolated BLOCKED_RULE conflict"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinGroup"
              }
            }
          }
        }
      }
    },
    "/groups/{id}/leave": {
      "post": {
        "operationId": "leaveGroupBooking",
        "tags": [
          "Group Booking"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Group"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Membership, owner, payer or signed invite mismatch"
          },
          "409": {
            "description": "Deadline, slot, allocation or isolated BLOCKED_RULE conflict"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/groups/{id}/cancel": {
      "post": {
        "operationId": "cancelUnpaidGroup",
        "tags": [
          "Group Booking"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Group"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Membership, owner, payer or signed invite mismatch"
          },
          "409": {
            "description": "Deadline, slot, allocation or isolated BLOCKED_RULE conflict"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelInput"
              }
            }
          }
        }
      }
    },
    "/groups/{id}/allocations": {
      "put": {
        "operationId": "splitGroupCost",
        "tags": [
          "Group Booking"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Group"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Membership, owner, payer or signed invite mismatch"
          },
          "409": {
            "description": "Deadline, slot, allocation or isolated BLOCKED_RULE conflict"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupSplit"
              }
            }
          }
        }
      }
    },
    "/groups/{id}/members/{memberId}/payable": {
      "get": {
        "operationId": "getContributionPayable",
        "tags": [
          "Group Booking"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Payable"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Membership, owner, payer or signed invite mismatch"
          },
          "409": {
            "description": "Deadline, slot, allocation or isolated BLOCKED_RULE conflict"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "memberId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Slot": {
        "type": "object",
        "properties": {
          "startsAt": {
            "type": "string",
            "format": "date-time"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time"
          },
          "state": {
            "type": "string",
            "enum": [
              "ELIGIBLE",
              "AVAILABLE",
              "BLOCKED"
            ]
          },
          "reason": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "minimum": 0,
            "exclusiveMinimum": true,
            "nullable": true
          },
          "currency": {
            "type": "string"
          },
          "ruleId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "ruleVersion": {
            "type": "integer"
          },
          "ruleLabel": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "startsAt",
          "endsAt",
          "state",
          "reason",
          "amount",
          "currency",
          "ruleId",
          "ruleVersion",
          "ruleLabel"
        ]
      },
      "Quote": {
        "type": "object",
        "properties": {
          "facilityId": {
            "type": "string",
            "format": "uuid"
          },
          "courtId": {
            "type": "string",
            "format": "uuid"
          },
          "startsAt": {
            "type": "string",
            "format": "date-time"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time"
          },
          "amount": {
            "type": "number",
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "currency": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          },
          "segments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Slot"
            }
          }
        },
        "required": [
          "facilityId",
          "courtId",
          "startsAt",
          "endsAt",
          "amount",
          "currency",
          "timezone",
          "segments"
        ]
      },
...
      "Booking": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "facilityId": {
            "type": "string",
            "format": "uuid"
          },
          "courtId": {
            "type": "string",
            "format": "uuid"
          },
          "customerId": {
            "type": "string",
            "format": "uuid"
          },
          "currentHolderId": {
            "type": "string",
            "format": "uuid"
          },
          "createdBy": {
            "type": "string",
            "format": "uuid"
          },
          "startsAt": {
            "type": "string",
            "format": "date-time"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "CONFIRMED",
              "CHECKED_IN",
              "COMPLETED",
              "CANCELLED",
              "EXPIRED"
            ]
          },
          "source": {
            "type": "string",
            "enum": [
              "ONLINE",
              "OFFLINE_COUNTER"
            ]
          },
          "guestName": {
            "type": "string",
            "nullable": true
          },
          "guestPhone": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "priceSnapshot": {
            "$ref": "#/components/schemas/Quote"
          },
          "holdExpiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "paymentId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "paidAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "checkedInAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "version": {
            "type": "integer"
          },
          "checkinTokenVersion": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "facilityId",
          "courtId",
          "customerId",
          "currentHolderId",
          "createdBy",
          "startsAt",
          "endsAt",
          "status",
          "source",
          "guestName",
          "guestPhone",
          "amount",
          "currency",
          "priceSnapshot",
          "holdExpiresAt",
          "paymentId",
          "paidAt",
          "checkedInAt",
          "completedAt",
          "version",
          "checkinTokenVersion"
        ]
      },
...
      "Payable": {
        "type": "object",
        "properties": {
          "bookingId": {
            "type": "string",
            "format": "uuid"
          },
          "payerId": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "purpose": {
            "type": "string"
          },
          "memberId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "required": [
          "bookingId",
          "payerId",
          "amount",
          "currency",
          "expiresAt",
          "purpose",
          "memberId"
        ]
      },
      "CancelInput": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          }
        },
        "required": [
          "reason"
        ]
      },
...
      "CreateGroup": {
        "type": "object",
        "required": [
          "holdId",
          "maxMembers"
        ],
        "properties": {
          "holdId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "maxLength": 180
          },
          "maxMembers": {
            "type": "integer",
            "minimum": 1,
            "maximum": 50
          }
        }
      },
      "JoinGroup": {
        "type": "object",
        "required": [
          "code"
        ],
        "properties": {
          "code": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          }
        }
      },
      "GroupMember": {
        "type": "object",
        "required": [
          "id",
          "userId",
          "displayName",
          "active",
          "amountDue",
          "amountPaid",
          "paymentState"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "amountDue": {
            "type": "number",
            "minimum": 0
          },
          "amountPaid": {
            "type": "number",
            "minimum": 0
          },
          "paymentState": {
            "type": "string",
            "enum": [
              "UNPAID",
              "WAITING_FOR_PAYMENT",
              "PAID",
              "PAID_BY_OWNER"
            ]
          },
          "paymentRequestedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "Group": {
        "type": "object",
        "required": [
          "id",
          "ownerId",
          "name",
          "state",
          "deadline",
          "maxMembers",
          "allocationsLocked",
          "booking",
          "members",
          "totalPaid"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "enum": [
              "GROUP_PENDING",
              "CONFIRMED",
              "GROUP_EXPIRED",
              "GROUP_CANCELLED"
            ]
          },
          "deadline": {
            "type": "string",
            "format": "date-time"
          },
          "maxMembers": {
            "type": "integer"
          },
          "allocationsLocked": {
            "type": "boolean"
          },
          "booking": {
            "$ref": "#/components/schemas/Booking"
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupMember"
            }
          },
          "totalPaid": {
            "type": "number",
            "minimum": 0
          }
        }
      },
      "GroupInvite": {
        "type": "object",
        "required": [
          "code",
          "path",
          "qrSvg",
          "expiresAt"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "qrSvg": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "qrPngDataUrl": {
            "description": "640px PNG data URL for downloading and scanning the invitation",
            "type": "string"
          }
        }
      },
      "GroupAllocation": {
        "type": "object",
        "required": [
          "memberId",
          "amount"
        ],
        "properties": {
          "memberId": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "type": "number",
            "minimum": 0,
            "multipleOf": 1
          }
        }
      },
      "GroupSplit": {
        "type": "object",
        "required": [
          "mode"
        ],
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "EQUAL",
              "CUSTOM"
            ]
          },
          "allocations": {
            "type": "array",
            "maxItems": 50,
            "items": {
              "$ref": "#/components/schemas/GroupAllocation"
            },
            "nullable": true
          }
        }
      },
...
    }
  },
...
}